+++ src/coverage_clients/jacoco.rs
@@ -56,7 +56,7 @@ pub struct JacocoCoverageClient<'a> {
     first_unused_idx: usize,
 
     stream: TeeStream,
-    max_ratio: (u64, u64),
+    max_ratio: (u32, u32),
     done: bool,
     latest_coverage_information: Vec<u8>,
     jacoco_dump_output_dir: Option<PathBuf>,
@@ -273,12 +273,12 @@ impl<'a> JacocoCoverageClient<'a> {
         })
     }
 
-    fn coverage_ratio(&mut self) -> (u64, u64) {
+    fn coverage_ratio(&mut self) -> (u32, u32) {
         let ones_count = self
             .cov_map_total
             .iter()
-            .fold(0u64, |sum, val| sum + u64::from(val.count_ones()));
-        let total_bits = self.first_unused_idx as u64 * 8;
+            .fold(0u32, |sum, val| sum + val.count_ones());
+        let total_bits = self.first_unused_idx as u32 * 8;
         (ones_count, total_bits)
     }
 
@@ -355,7 +355,7 @@ impl CoverageClient for JacocoCoverageClient<'_> {
         self.cov_map.as_mut_ptr()
     }
 
-    fn max_coverage_ratio(&mut self) -> (u64, u64) {
+    fn max_coverage_ratio(&mut self) -> (u32, u32) {
         let (count, total) = self.coverage_ratio();
         // update the max coverage ratio
         self.max_ratio.0 = std::cmp::max(self.max_ratio.0, count);
